Split the metadata update command line options in two
The 'update' feature is built on top of a general system for syncing changes. That
system is mostly unused now (except for servo, perhaps), and it's confusing to have
all the extra functionality documented in the parts that are only used for updating
metadata files based on existing results.

To ensure backward compatibility with existing consumers, the previous function
names are preserved with the full set of functionality.