Use separate transform tree nodes for translate/rotate/scale/offset/transform.
Rather than using a single node in the transform tree, use separate
nodes for the CSS 'translate', 'rotate', 'scale', 'offset-*', and
'transform' properties.

This is needed to animate these properties simultaneously on the
compositor.

It only increases the number of nodes when more than one of these
features are in use on the same element.

https://docs.google.com/document/d/18Kf6gwMX3-4D8eLBcGFn2fQpxG31vkDcq1nESFQHB9U/edit
documents a few of the design decisions that went into this change.

This makes the following minor behavior change:

 * This distinguishes will-change: scale, will-change: rotate,
   will-change: translate from the other properties (transform,
   transform-style, perspective) that are transform-related compositing
   triggers and ensure creation of transform nodes, since they now
   ensure the creation *different* transform nodes.  In other words,
   they now trigger a slightly different optimization rather than
   exactly the same one.  Likewise, will-change: offset-path and
   will-change: offset-position continue to force compositing but no
   longer have any effect on the cc transform tree.

This intermediate state of bug 696374 only being partly fixed also
introduces a regression when animations of rotate, translate, or scale
(properties that are not yet shipped) run on the compositor on an
element that also has a CC transform tree node generated for the
transform property.  This failure is shown by:

 * The existing test
   virtual/threaded-no-composited-antialiasing/animations/timing/translate-neutral-keyframe-easing.html

 * The newly added test
   external/wpt/css/css-transforms/animation/rotate-animation-with-will-change-transform-001.html
   in both virtual/threaded/ and virtual/composite-relative-keyframes/

For the former test, this change creates a new
animations/timing/translate-neutral-keyframe-easing-2.html that doesn't
exhibit this regression.

(I'm not particularly happy about some of the code duplication in this
change, but improving the situation also seems likely to be an even
*larger* change.)

Bug: 696374
Change-Id: I21ea753d14216d46fddcebc253b6a196525ad55e