Remove <link>-based APIs from subresource loading for Web Bundles
This removes the support of the <link>-based API, which is no longer
supported.

Most tests of <link>-based API were already removed at the preceding
CLs:
- https://crrev.com/c/3546136
- https://crrev.com/c/3539725

This CL removes the remaining tests or migrated them into tests for
<script>-based API. Hopefully, we found all tests and there is no
remaining tests.

Regarding implementation, a further optional clean up could be possible,
but that can be done in a follow-up CL.

Bug: 1306725
Change-Id: I75a1bf3ff42942e029af4a0690da32a69fcfaea3