[basic-card] Migrate payment-handler WPTs
With basic card now disabled and soon to be removed, migrate the
payment-handler web platform tests to use URL based payment methods
instead of basic-card.

The test cases that are exclusively related to basic-card are removed.
Many tests are updated to use a simple payment app with just a
paymentrequest event listener, app-simple.js, which is migrated from
basic-card.js, with the basic-card related values removed.

The untrusted-event tests are updated to use an example URL method as
they don't require a valid payment method.

Bug: 1209835
Change-Id: I79b95ff7df30eda7cbf253fa895566e6e2214aa0